pyproject.toml
[tool.poetry]
name = "nonebot-plugin-sqlalchemy"
version = "0.2.2"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
readme = "README.MD"
license = "MIT"
repository = "https://github.com/ssttkkl/nonebot-plugin-sqlalchemy"
packages = [
    { include = "nonebot_plugin_sqlalchemy", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }

[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.17.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
